Refactor `WorkspaceEndpoints.cs` to follow SRP.
Split the monolithic `WorkspaceEndpoints.cs` file into three distinct endpoint classes based on logical domain boundaries: `WorkspaceEndpoints`, `WorkspaceMemberEndpoints`, and `WorkspacePropertyEndpoints`.

This improves maintainability and aligns with the Minimal API Routing Standard without altering existing functionality or breaking backward compatibility. Updates were correctly registered in `Program.cs`.
