+++ backend/Valora.Api/Endpoints/WorkspaceEndpoints.cs
@@ -20,26 +20,6 @@ public static class WorkspaceEndpoints
         group.MapGet("/", GetUserWorkspaces);
         group.MapGet("/{id}", GetWorkspace);
         group.MapDelete("/{id}", DeleteWorkspace);
-
-        group.MapGet("/{id}/members", GetMembers);
-        group.MapPost("/{id}/members", InviteMember)
-            .AddEndpointFilter<Valora.Api.Filters.ValidationFilter<InviteMemberDto>>();
-
-        group.MapDelete("/{id}/members/{memberId}", RemoveMember);
-
-        group.MapGet("/{id}/properties", GetSavedProperties);
-        group.MapPost("/{id}/properties", SaveProperty)
-            .AddEndpointFilter<Valora.Api.Filters.ValidationFilter<SavePropertyDto>>();
-
-        group.MapPost("/{id}/properties/from-report", SavePropertyFromReport)
-            .AddEndpointFilter<Valora.Api.Filters.ValidationFilter<SavePropertyFromReportDto>>();
-
-        group.MapDelete("/{id}/properties/{savedPropertyId}", RemoveSavedProperty);
-
-        group.MapGet("/{id}/properties/{savedPropertyId}/comments", GetComments);
-        group.MapPost("/{id}/properties/{savedPropertyId}/comments", AddComment)
-            .AddEndpointFilter<Valora.Api.Filters.ValidationFilter<AddCommentDto>>();
-
         group.MapGet("/{id}/activity", GetActivityLogs);
 
         return group;
@@ -106,141 +86,6 @@ public static class WorkspaceEndpoints
         return Results.NoContent();
     }
 
-    private static async Task<IResult> GetMembers(
-        ClaimsPrincipal user,
-        Guid id,
-        IWorkspaceMemberService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        var result = await service.GetMembersAsync(userId, id, ct);
-        return Results.Ok(result);
-    }
-
-    private static async Task<IResult> InviteMember(
-        ClaimsPrincipal user,
-        Guid id,
-        [FromBody] InviteMemberDto dto,
-        IWorkspaceMemberService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        await service.AddMemberAsync(userId, id, dto, ct);
-        return Results.Ok();
-    }
-
-    private static async Task<IResult> RemoveMember(
-        ClaimsPrincipal user,
-        Guid id,
-        Guid memberId,
-        IWorkspaceMemberService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        await service.RemoveMemberAsync(userId, id, memberId, ct);
-        return Results.NoContent();
-    }
-
-    private static async Task<IResult> GetSavedProperties(
-        ClaimsPrincipal user,
-        Guid id,
-        IWorkspacePropertyService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        var result = await service.GetSavedPropertiesAsync(userId, id, ct);
-        return Results.Ok(result);
-    }
-
-    private static async Task<IResult> SaveProperty(
-        ClaimsPrincipal user,
-        Guid id,
-        [FromBody] SavePropertyDto request,
-        IWorkspacePropertyService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        var result = await service.SavePropertyAsync(userId, id, request.PropertyId, request.Notes, ct);
-        return Results.Ok(result);
-    }
-
-    /// <summary>
-    /// Saves a newly generated Context Report directly to a Workspace.
-    /// </summary>
-    /// <remarks>
-    /// <para>
-    /// <strong>Why save reports?</strong> Context reports are dynamically generated in real-time ("Fan-Out").
-    /// Saving a report to a workspace takes a "snapshot" of the data and persists it to the database,
-    /// allowing users to revisit the exact stats, share them with team members, and add comments.
-    /// </para>
-    /// </remarks>
-    private static async Task<IResult> SavePropertyFromReport(
-        ClaimsPrincipal user,
-        Guid id,
-        [FromBody] SavePropertyFromReportDto request,
-        IWorkspacePropertyService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        var result = await service.SaveContextReportAsync(userId, id, request.Report, request.Notes, ct);
-        return Results.Ok(result);
-    }
-
-    private static async Task<IResult> RemoveSavedProperty(
-        ClaimsPrincipal user,
-        Guid id,
-        Guid savedPropertyId,
-        IWorkspacePropertyService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        await service.RemoveSavedPropertyAsync(userId, id, savedPropertyId, ct);
-        return Results.NoContent();
-    }
-
-    private static async Task<IResult> GetComments(
-        ClaimsPrincipal user,
-        Guid id,
-        Guid savedPropertyId,
-        IWorkspacePropertyService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        var result = await service.GetCommentsAsync(userId, id, savedPropertyId, ct);
-        return Results.Ok(result);
-    }
-
-    private static async Task<IResult> AddComment(
-        ClaimsPrincipal user,
-        Guid id,
-        Guid savedPropertyId,
-        [FromBody] AddCommentDto dto,
-        IWorkspacePropertyService service,
-        CancellationToken ct)
-    {
-        var userId = user.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (userId == null) return Results.Unauthorized();
-
-        var result = await service.AddCommentAsync(userId, id, savedPropertyId, dto, ct);
-        return Results.Ok(result);
-    }
-
     private static async Task<IResult> GetActivityLogs(
         ClaimsPrincipal user,
         Guid id,
@@ -253,5 +98,4 @@ public static class WorkspaceEndpoints
         var result = await service.GetActivityLogsAsync(userId, id, ct);
         return Results.Ok(result);
     }
-
 }
